requirements.txt
python-socketio
flask-socketio
google-cloud-speech